+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- verifyJWT to call API instance of using cognito
+- verifyJWT to call API instead of using cognito
 
 ## [0.19.16] - 2022-11-02